+++ frontend/src/App.js
@@ -7,7 +7,7 @@ import { ConnectedRouter } from "connected-react-router";
 
 
 import {
-    Main, ReviewDetail, ReviewCreate, ReviewEdit, PaperDetail,
+    Main, ReviewDetail, PaperDetail, ReviewControl,
 } from "./containers";
 
 function App(props) {
@@ -16,9 +16,25 @@ function App(props) {
             <div className="App">
                 <Switch>
                     <Route path="/main" exact component={Main} />
-                    <Route path="/papers/:paper_id/create" exact component={ReviewCreate} />
+                    <Route
+                      path="/papers/:paper_id/create"
+                      exact
+                      render={() => (
+                          <div>
+                              <ReviewControl mode={0} />
+                          </div>
+                      )}
+                    />
                     <Route path="/papers/:paper_id/:review_id" exact component={ReviewDetail} />
-                    <Route path="/papers/:paper_id/:review_id/edit" exact component={ReviewEdit} />
+                    <Route
+                      path="/papers/:paper_id/:review_id/edit"
+                      exact
+                      render={() => (
+                          <div>
+                              <ReviewControl mode={1} />
+                          </div>
+                      )}
+                    />
                     <Route path="/papers/:id" exact component={PaperDetail} />
                 </Switch>
             </div>
