--- frontend/src/containers/Review/ReviewCreate/ReviewCreate.css
@@ -1,48 +1,0 @@
-.review-create .review-create-page{
-    background: rgb(242,243,247);
-    margin-left: 18%;
-    display: flex;
-    min-height: 1000px;
-}
-
-.review-create .review-create-page .board{
-    width: 80%;
-    margin-top: 5%;
-    margin-left: 5%;
-    margin-right: 10%;
-    background: white;
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    flex-direction: column;
-}
-
-.review-create .review-create-page .board .paper-spec{
-    width: 80%;
-    background: white;
-    display: flex;
-    justify-content: flex-start;
-    flex-direction: column;
-    margin-bottom: 3%;
-    border-radius: 0.5em;
-    border-bottom: 2px solid  rgb(242,243,247);
-} 
-
-.review-create .board .form-title{
-    width: 80%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    font-size: 120%;
-    font-weight: bold;
-}
-
-.review-create .board .form-content{
-    margin-top: 5%;
-    width: 80%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    font-size: 120%;
-    font-weight: bold;
-}
